+++ src/components/debug/debug.css
