--- src/components/general/button/button.js
@@ -1,11 +1,0 @@
-// primary
-
-// secondary
-
-// warning
-
-// danger
-
-// success
-
-// info
