+++ src/index.ts
@@ -10,9 +10,9 @@ import { Playlist, iRealRenderer } from "ireal-renderer";
 // import Annotation from './Annotation'
 import { Controls, transposeService } from "./Control";
 import { SetlistNav } from "./Setlist";
-import { DebugNavContent, Tracer } from "./Debug";
+import { DebugNavContent, Tracer } from "./components/debug/Debug";
 import { State } from "./State";
-import { Nav, InactiveNavControl } from "./components/leo/ui/navigation/nav";
+import { Nav, InactiveNavControl } from "./components/navigation/nav";
 import "./styles/screens.css";
 
 export const playlist = new Playlist(ireal);
